Fix packet parsing correctness and robustness findings (#552)
## Summary

- Replace dead-code try/catch platform validation and the
exception-driven `AdjustSessionType` fallback with explicit
`Enum.IsDefined` / `Unknown` fallbacks, matching the existing
`TyreCompoundMapper` pattern; cut `DriverName` at the first null
terminator instead of trimming the full fixed-size buffer.
- Widen several inconsistently typed spec `uint8` fields
(`PitLimiterStatus`, `CarIndex`/`TeamId` on the time trial data set,
`NumberOfLaps`/`LapStartIndex`, `TimeOffset`) to `ushort` to match the
rest of the codebase.
- Retype `TrackId` from `ushort` to `short` across all game years so the
spec's `-1 = unknown` sentinel is representable, and fix the raw-byte
read to reinterpret it as signed.
- Pre-populate the per-car/per-lap array elements in the
`FinalClassificationData`, `LapDataComplete`, `ParticipantsData`, and
`SessionHistoryData` constructors (2022-2025), matching the existing
`CarStatus`/`CarTelemetry` pattern, so consumers never observe `null`
entries before the parser runs.
- Give `EventCode`/`DriverName` explicit `string.Empty` defaults on the
2023-2025 event/participant data classes.
- Clarify the `LapData2024` XML docs for
`DeltaToCarInFront`/`DeltaToRaceLeader` to state they hold only the
milliseconds part, pointing to the sibling `*Minutes` properties.
- Retype `TractionControl`/`GearboxAssist` to `bool` on
`ITimeTrialDataSetBase`, since the TimeTrial packet spec documents these
two fields as binary assist flags for every game year, unlike the
multi-valued enums used on car status/setup data. All time trial data
set classes (2024-2026) now consistently expose them as `bool` through
the shared base interface.

## Linked issues

Closes #80, Closes #81, Closes #89, Closes #63, Closes #100, Closes
#103, Closes #106, Closes #111, Closes #112

## Review notes

- `ITimeTrialDataSetBase` now declares `TractionControl`/`GearboxAssist`
as `bool` directly, since the TimeTrial spec byte is binary for every
game year (2024-2026); the concrete classes need no per-year override,
and the shared parser assigns the boolean uniformly across years.
- The `TrackId` retype ripples into `SessionData2019`-`SessionData2023`
(2024-2026 inherit from `SessionData2023`); all downstream consumers
(`SessionProcessor`, `TrackRepositoryCache`, DB entities) already use
wider integer types, so no further changes were required there.

+++ F1Server.Core/Packets/Data/F1-2021/CarStatusData2021.cs
@@ -33,7 +33,7 @@ public class CarStatusData2021 : ICarStatusData2021
     /// <summary>
     /// Pit limiter status - 0 - off - 1 - on
     /// </summary>
-    public uint PitLimiterStatus { get; set; }
+    public ushort PitLimiterStatus { get; set; }
 
     /// <summary>
     /// Current fuel mass
